Use ATS for smoke test
